requirements.txt
Pillow==9.0.1
PyMuPDF==1.18.6
Unidecode==1.1.2
greenlet==1.1.1
SQLAlchemy==1.4.23
psycopg2==2.9.1
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
spacy==3.1.3
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
murmurhash==1.0.7
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.8
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
smart-open==5.2.1
spacy-legacy==3.0.9
srsly==2.4.3
thinc==8.0.15
threadpoolctl==3.1.0
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
